% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_mass.R
\name{sequence_mass}
\alias{sequence_mass}
\title{Calculate the mass of a poly amino acid}
\usage{
sequence_mass(sequence = NULL, charge = 0)
}
\arguments{
\item{sequence}{as character string}

\item{charge}{an integer}
}
\value{
a vector
}
\description{
\code{sequence_mass()} get the mass of a poly amino acid
}
